refactor(editor): separate line number logic to util
