+++ app/src/main/java/com/csapp/sqli/utils/LineNumberManager.kt
@@ -1,5 +1,26 @@
 package com.csapp.sqli.utils
 
-class LineNumberManager {
-    
+import android.text.Layout
+import com.csapp.sqli.model.LineNumber
+
+object LineNumberManager {
+    private val modelLineNumber = LineNumber()
+
+    // If line number changed, update model and render line number text view
+    fun updateLineNumbers(it: Layout): String {
+        if (it.lineCount != modelLineNumber.number) {
+            modelLineNumber.content = generateLineNumber(it.lineCount)
+            modelLineNumber.number = it.lineCount
+        }
+        return modelLineNumber.content
+    }
+
+    // Generate line number like ""1/n2/n3/n ... count"
+    private fun generateLineNumber(count: Int): String {
+        return buildString {
+            for (i in 1..count) {
+                append("$i\n")
+            }
+        }
+    }
 }
