+++ app/src/main/java/com/csapp/sqli/viewmodel/EditorViewModel.kt
@@ -1,16 +1,13 @@
 package com.csapp.sqli.viewmodel
 
-import android.text.Layout
 import android.widget.EditText
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.csapp.sqli.model.LineNumber
 import com.csapp.sqli.repository.DatabaseRepository
+import com.csapp.sqli.utils.LineNumberManager
 
 class EditorViewModel(private val databaseRepository: DatabaseRepository) : ViewModel() {
-    private val modelLineNumber = LineNumber()
-
     private val _editTextStatement = MutableLiveData<EditText>()
     val editTextStatement: MutableLiveData<EditText> = _editTextStatement
 
@@ -19,16 +16,8 @@ class EditorViewModel(private val databaseRepository: DatabaseRepository) : View
 
     // Data binding with Activity_editor.xml #afterTextChanged()
     fun onStatementChanged() {
-        val editText = editTextStatement.value
-        editText?.layout?.let { updateLineNumbers(it) }
-    }
-
-    // If line number changed, update model and render line number text view
-    private fun updateLineNumbers(it: Layout) {
-        if (it.lineCount != modelLineNumber.number) {
-            modelLineNumber.content = generateLineNumber(it.lineCount)
-            modelLineNumber.number = it.lineCount
-            _textViewLineNumber.value = modelLineNumber.content
+        editTextStatement.value?.layout?.let {
+            _textViewLineNumber.value = LineNumberManager.updateLineNumbers(it)
         }
     }
 
@@ -46,14 +35,5 @@ class EditorViewModel(private val databaseRepository: DatabaseRepository) : View
     // Check if a statement starts with "SELECT" ignore case
     private fun isSelectStatement(statement: String): Boolean {
         return statement.startsWith("SELECT", ignoreCase = true)
-    }
-
-    // Generate line number like ""1/n2/n3/n ... count"
-    private fun generateLineNumber(count: Int): String {
-        val stringBuilder = StringBuilder()
-        for (i in 1..count) {
-            stringBuilder.append("$i\n")
-        }
-        return stringBuilder.toString()
     }
 }
